chore: refactor test providers
