+++ crates/providers-alloy/src/pipeline.rs
@@ -66,3 +66,45 @@ pub fn new_online_pipeline(
         .origin(origin)
         .build()
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::OnlineBlobProvider;
+    use kona_derive::prelude::OriginProvider;
+
+    #[test]
+    fn test_new_online_pipeline() {
+        let rollup_config = Arc::new(RollupConfig::default());
+        let chain_provider =
+            AlloyChainProvider::new_http("http://127.0.0.1:8545".try_into().unwrap());
+        let l2_chain_provider = AlloyL2ChainProvider::new_http(
+            "http://127.0.0.1:9545".try_into().unwrap(),
+            rollup_config.clone(),
+        );
+        let beacon_client = OnlineBeaconClient::new_http("http://127.0.0.1:5555".into());
+        let blob_provider: OnlineBlobProvider<_, SimpleSlotDerivation> =
+            OnlineBlobProvider::new(beacon_client, None, None);
+        let blob_provider = OnlineBlobProviderWithFallback::new(blob_provider, None);
+        let dap_source =
+            EthereumDataSource::new(chain_provider.clone(), blob_provider, &rollup_config);
+        let builder = StatefulAttributesBuilder::new(
+            rollup_config.clone(),
+            l2_chain_provider.clone(),
+            chain_provider.clone(),
+        );
+        let origin = BlockInfo::default();
+
+        let pipeline = new_online_pipeline(
+            rollup_config.clone(),
+            chain_provider,
+            dap_source,
+            l2_chain_provider,
+            builder,
+            origin,
+        );
+
+        assert_eq!(pipeline.rollup_config, rollup_config);
+        assert_eq!(pipeline.origin(), Some(origin));
+    }
+}
